# Conflicts:
#	dao-impl/src/main/java/jm/dao/ChannelDAOImpl.java

+++ service-impl/src/main/java/jm/ChannelServiceImpl.java
@@ -41,7 +41,7 @@ public class ChannelServiceImpl implements ChannelService {
     }
 
     @Override
-    public Channel getChannelById(int id) {
+    public Channel getChannelById(Long id) {
         return channelDAO.getChannelById(id);
     }
